+++ src/AceCommon.h
@@ -32,8 +32,8 @@ This is the Doxygen documentation for the
 #ifndef ACE_COMMON_H
 #define ACE_COMMON_H
 
-#include "ace_common/arithmetic.h"
-#include "ace_common/pstrings.h"
+#include "arithmetic/arithmetic.h"
+#include "pstrings/pstrings.h"
 #include "print_str/PrintStr.h"
 #include "print_utils/printPadTo.h"
 #include "print_utils/printfTo.h"
